minor fixes in part1, part2 source + fixed redundant includes + fixed unsigned comparison * modified part2 README

+++ part1/src/main.cpp
@@ -1,15 +1,10 @@
 #include <string>
-
 #include <ncurses.h>
-
-// #include <quiz.h>
 
 
 
 int main(int argc, char **argv){
-    // parse 
-    // quiz q;
-    // quiz();
+    
     WINDOW *wnd;
     wnd = initscr();
     cbreak();
@@ -20,7 +15,7 @@ int main(int argc, char **argv){
     move(5, 5);
 
     std::string text = "Hello world!";
-    for(int i = 0; i < text.size(); i++) {
+    for(size_t i = 0; i < text.size(); i++) {
         addch(text[i]);
         addch(' ');
     }
